requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic-settings
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
requests
psycopg2-binary
